test(suite): modularize domains and add perf regression suite
Split the monolithic tests/crdt.test.ts into domain-focused suites (state-core, patch-diff-doc, merge-compaction, replica-session) and move shared helpers into tests/test-utils.ts for discoverability and maintenance.

Add tests/perf-regression.test.ts for hotspot regressions and add area-scoped package scripts (test:state-core, test:patch-diff-doc, test:merge-compaction, test:replica-session, test:perf-regression) to speed targeted runs.

Update README with targeted testing commands and include a patch changeset for the test architecture refactor.

Refs #22
